Launch notes
the app is not having the python backend fills and the buttons are disabled, and the graphical interface is currently launched in electron, install nodejs, and electron, on your pc, then run electron from the terminal/ cmd. and navigate to the project folder, then launch main.js file. i hope to use GTK+ for the GUI in the final application after i get the time to embark on the python backend files, for weather, pest, and disease detection. am willing to accept any collaboration for the AI-ML python dev.